+++ src/kalinka_plugin_qobuz/module_setup.py
@@ -20,7 +20,7 @@ logger = logging.getLogger(__name__.split(".")[-1])
 
 
 class KalinkaPluginQobuz(InputModulePlugin):
-    REQUIRES_SDK = ">=1,<2"
+    REQUIRES_SDK = ">=2,<3"
     PLUGIN_ID = "qobuz"
     CONFIG_MODEL = QobuzConfig